Refactor notification scripts into modular folder
