--- scripts/notifyUsers.ts
@@ -1,1192 +1,0 @@
-import "dotenv/config";
-import { mongodbConnect } from "../db.ts";
-import { JORFSearchItem } from "../entities/JORFSearchResponse.ts";
-import { FunctionTags } from "../entities/FunctionTags.ts";
-import {
-  IOrganisation,
-  IPeople,
-  IUser,
-  MessageApp,
-  WikidataId
-} from "../types.ts";
-import People from "../models/People.ts";
-import User from "../models/User.ts";
-import { Types } from "mongoose";
-import umami from "../utils/umami.ts";
-import { JORFtoDate } from "../utils/date.utils.ts";
-import { formatSearchResult } from "../utils/formatSearchResult.ts";
-import {
-  callJORFSearchDay,
-  cleanPeopleName,
-  getJORFSearchLinkFunctionTag,
-  getJORFSearchLinkOrganisation,
-  getJORFSearchLinkPeople
-} from "../utils/JORFSearch.utils.ts";
-import Organisation from "../models/Organisation.ts";
-import { ExternalMessageOptions, sendMessage } from "../entities/Session.ts";
-import { WhatsAppAPI } from "whatsapp-api-js/middleware/express";
-import { ErrorMessages } from "../entities/ErrorMessages.ts";
-import { WHATSAPP_API_VERSION } from "../entities/WhatsAppSession.ts";
-import { SignalCli } from "signal-sdk";
-import {
-  NotificationTask,
-  dispatchTasksToMessageApps
-} from "../utils/notificationDispatch.ts";
-/*
-import {
-  MatrixClient,
-  RustSdkCryptoStorageProvider,
-  SimpleFsStorageProvider
-} from "matrix-bot-sdk";
-*/
-import { sendMainMenu } from "../commands/default.ts";
-
-// Number of days to go back: 0 means we just fetch today's info
-const SHIFT_DAYS = 30;
-
-const { ENABLED_APPS } = process.env;
-if (ENABLED_APPS === undefined) throw new Error("ENABLED_APPS env var not set");
-const enabledApps = JSON.parse(ENABLED_APPS) as MessageApp[];
-
-const invalidApps: string[] = [];
-for (const app of enabledApps) {
-  if (!["Telegram", "Matrix", "WhatsApp", "Signal"].includes(app))
-    invalidApps.push(app);
-}
-if (invalidApps.length > 0)
-  throw new Error(
-    `Invalid message app${invalidApps.length > 0 ? "s" : ""}: ${invalidApps.join(", ")}`
-  );
-
-/*
-let matrixClient: MatrixClient | undefined = undefined;
-if (enabledApps.includes("Matrix")) {
-  const { MATRIX_HOME_URL, MATRIX_BOT_TOKEN } = process.env;
-  if (MATRIX_HOME_URL == undefined || MATRIX_BOT_TOKEN == undefined)
-    throw new Error("MATRIX env is not set");
-
-  const storageProvider = new SimpleFsStorageProvider("matrix-bot.json");
-  const cryptoProvider = new RustSdkCryptoStorageProvider("matrix-crypto");
-
-  matrixClient = new MatrixClient(
-    "https://" + MATRIX_HOME_URL,
-    MATRIX_BOT_TOKEN,
-    storageProvider,
-    cryptoProvider
-  );
-}
-*/
-
-let whatsAppAPI: WhatsAppAPI | undefined = undefined;
-if (enabledApps.includes("WhatsApp")) {
-  const { WHATSAPP_USER_TOKEN, WHATSAPP_APP_SECRET, WHATSAPP_VERIFY_TOKEN } =
-    process.env;
-  if (
-    WHATSAPP_USER_TOKEN === undefined ||
-    WHATSAPP_APP_SECRET === undefined ||
-    WHATSAPP_VERIFY_TOKEN === undefined
-  )
-    throw new Error(ErrorMessages.WHATSAPP_ENV_NOT_SET);
-
-  whatsAppAPI = new WhatsAppAPI({
-    token: WHATSAPP_USER_TOKEN,
-    appSecret: WHATSAPP_APP_SECRET,
-    webhookVerifyToken: WHATSAPP_VERIFY_TOKEN,
-    v: WHATSAPP_API_VERSION
-  });
-}
-
-let signalCli: SignalCli | undefined = undefined;
-if (enabledApps.includes("Signal")) {
-  const { SIGNAL_BAT_PATH, SIGNAL_PHONE_NUMBER } = process.env;
-  if (SIGNAL_BAT_PATH === undefined || SIGNAL_PHONE_NUMBER === undefined)
-    throw new Error(ErrorMessages.SIGNAL_ENV_NOT_SET);
-
-  signalCli = new SignalCli(SIGNAL_BAT_PATH, SIGNAL_PHONE_NUMBER);
-  await signalCli.connect();
-}
-
-const messageAppsOptions: ExternalMessageOptions = {
-  //matrixClient,
-  signalCli: signalCli,
-  whatsAppAPI: whatsAppAPI
-};
-
-type GroupIdentifier = string | string[] | null | undefined;
-
-interface NotificationGroupingConfig {
-  getGroupId: (record: JORFSearchItem) => GroupIdentifier;
-  fallbackLabel?: string;
-  formatGroupTitle?: (options: {
-    groupId: string;
-    markdownLinkEnabled: boolean;
-  }) => string;
-  sortGroupIds?: (groupIds: string[]) => string[];
-  omitOrganisationNames?: boolean;
-}
-
-const DEFAULT_GROUP_SEPARATOR = "====================\n\n";
-const DEFAULT_SUBGROUP_SEPARATOR = "--------------------\n\n";
-
-function normaliseGroupId(id: string | null | undefined): string | null {
-  if (id === undefined || id === null) return null;
-  const trimmed = String(id).trim();
-  return trimmed.length > 0 ? trimmed : null;
-}
-
-function groupRecordsBy(
-  records: JORFSearchItem[],
-  config: NotificationGroupingConfig
-): Map<string, JORFSearchItem[]> {
-  const grouped = new Map<string, JORFSearchItem[]>();
-
-  for (const record of records) {
-    const rawGroupId = config.getGroupId(record);
-    const candidateIds = Array.isArray(rawGroupId)
-      ? rawGroupId
-      : [rawGroupId];
-
-    const validIds = candidateIds
-      .map((value) => normaliseGroupId(value))
-      .filter((value): value is string => value !== null);
-
-    const fallbackKey = normaliseGroupId(config.fallbackLabel);
-    const keysToUse =
-      validIds.length > 0
-        ? validIds
-        : fallbackKey
-        ? [fallbackKey]
-        : [];
-
-    for (const key of keysToUse) {
-      const existing = grouped.get(key) ?? [];
-      existing.push(record);
-      grouped.set(key, existing);
-    }
-  }
-
-  return grouped;
-}
-
-function orderGroupedEntries(
-  groupedMap: Map<string, JORFSearchItem[]>,
-  sort?: (groupIds: string[]) => string[]
-): Array<[string, JORFSearchItem[]]> {
-  const groupIds = sort ? sort([...groupedMap.keys()]) : [...groupedMap.keys()];
-  return groupIds.map((groupId) => [groupId, groupedMap.get(groupId) ?? []]);
-}
-
-function createFieldGrouping(
-  accessor: (record: JORFSearchItem) => GroupIdentifier,
-  options?: Omit<NotificationGroupingConfig, "getGroupId">
-): NotificationGroupingConfig {
-  return {
-    getGroupId: accessor,
-    fallbackLabel: options?.fallbackLabel,
-    formatGroupTitle: options?.formatGroupTitle,
-    sortGroupIds: options?.sortGroupIds,
-    omitOrganisationNames: options?.omitOrganisationNames
-  };
-}
-
-const CABINET_GROUP_FALLBACK_LABEL = "Autres ministères";
-
-const functionTagGroupingStrategies: Partial<
-  Record<FunctionTags, NotificationGroupingConfig>
-> = {
-  [FunctionTags["Cabinet ministériel"]]: createFieldGrouping(
-    (record) => record.ministere ?? record.cabinet,
-    {
-      fallbackLabel: CABINET_GROUP_FALLBACK_LABEL,
-      formatGroupTitle: ({ groupId }) => `🏛️ Ministère : *${groupId}*\n\n`,
-      sortGroupIds: (groupIds) => {
-        const withoutFallback = groupIds.filter(
-          (groupId) => groupId !== CABINET_GROUP_FALLBACK_LABEL
-        );
-        const sorted = withoutFallback.sort((a, b) =>
-          a.localeCompare(b, "fr", { sensitivity: "base" })
-        );
-        if (groupIds.includes(CABINET_GROUP_FALLBACK_LABEL))
-          sorted.push(CABINET_GROUP_FALLBACK_LABEL);
-        return sorted;
-      }
-    }
-  )
-};
-
-async function getJORFRecordsFromDate(
-  startDate: Date
-): Promise<JORFSearchItem[]> {
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  startDate.setHours(0, 0, 0, 0);
-
-  // Build the list of days to fetch (inclusive)
-  const dayCount = (today.getTime() - startDate.getTime()) / 86_400_000 + 1; // 1 day = 86 400 000 ms
-  const days: Date[] = Array.from({ length: dayCount }, (_, i) => {
-    const d = new Date(today);
-    d.setDate(today.getDate() - i);
-    return d;
-  });
-
-  // Fetch them concurrently (limit to e.g. 8 at a time to stay polite)
-  const limit = 8;
-  const chunks: Date[][] = [];
-  for (let i = 0; i < days.length; i += limit)
-    chunks.push(days.slice(i, i + limit));
-
-  const results: JORFSearchItem[][] = [];
-  for (const sub of chunks) {
-    results.push(...(await Promise.all(sub.map(callJORFSearchDay))));
-  }
-
-  return results
-    .flat()
-    .sort(
-      (a, b) =>
-        JORFtoDate(a.source_date).getTime() -
-        JORFtoDate(b.source_date).getTime()
-    );
-}
-
-// There is currently no way to check if a user has been notified of a tag update
-// Resuming an update thus require to force-notify users for all tags updates over the period.
-export async function notifyFunctionTagsUpdates(
-  updatedRecords: JORFSearchItem[]
-) {
-  if (updatedRecords.length == 0) return;
-
-  const functionTagValues: string[] = Object.values(FunctionTags);
-
-  // Initialize an empty tag map to store the categorized records
-  const updatedTagMap = new Map<FunctionTags, JORFSearchItem[]>();
-
-  // Build the tag map from the updated records
-  updatedRecords.forEach((item) => {
-    // Iterate through each key in the current record
-    (Object.keys(item) as (keyof JORFSearchItem)[]).forEach((key) => {
-      // Check if the key is a valid function tag
-      if (functionTagValues.includes(key)) {
-        const keyFctTag = key as FunctionTags;
-
-        // Update the tag map
-        const currentItems = updatedTagMap.get(keyFctTag) ?? [];
-        updatedTagMap.set(keyFctTag, [...currentItems, item]);
-      }
-    });
-  });
-
-  // Create a set of unique tag keys
-  const updatedTagSet = new Set<FunctionTags>(updatedTagMap.keys());
-
-  const usersFollowingTags: IUser[] = await User.find(
-    {
-      followedFunctions: {
-        $exists: true,
-        $not: { $size: 0 },
-        $elemMatch: {
-          functionTag: { $in: [...updatedTagSet] }
-        }
-      },
-      status: "active",
-      messageApp: { $in: enabledApps }
-    },
-    {
-      _id: 1,
-      messageApp: 1,
-      chatId: 1,
-      followedFunctions: { functionTag: 1, lastUpdate: 1 },
-      schemaVersion: 1
-    }
-  ).lean();
-  if (usersFollowingTags.length == 0) return;
-
-  const now = new Date();
-
-  const userUpdateTasks: NotificationTask<FunctionTags>[] = [];
-
-  for (const user of usersFollowingTags) {
-    const newUserTagsUpdates = new Map<FunctionTags, JORFSearchItem[]>();
-
-    user.followedFunctions
-      .filter((tagFollow) => updatedTagMap.has(tagFollow.functionTag))
-      .forEach((tagFollow) => {
-        const dateFilteredUserTagUpdates: JORFSearchItem[] = (
-          updatedTagMap.get(tagFollow.functionTag) ?? []
-        ).filter(
-          (record: JORFSearchItem) =>
-            JORFtoDate(record.source_date).getTime() >
-            tagFollow.lastUpdate.getTime()
-        );
-        if (dateFilteredUserTagUpdates.length > 0)
-          newUserTagsUpdates.set(
-            tagFollow.functionTag,
-            dateFilteredUserTagUpdates
-          );
-      });
-
-    // Calculate the total number of JORFSearchItem in the map
-    let totalUserRecordsCount = 0;
-    newUserTagsUpdates.forEach((items) => {
-      totalUserRecordsCount += items.length;
-    });
-
-    if (totalUserRecordsCount > 0)
-      userUpdateTasks.push({
-        userId: user._id,
-        messageApp: user.messageApp,
-        chatId: user.chatId,
-        updatedRecordsMap: newUserTagsUpdates,
-        recordCount: totalUserRecordsCount
-      });
-  }
-
-  if (userUpdateTasks.length == 0) return;
-
-  await dispatchTasksToMessageApps<FunctionTags>(
-    userUpdateTasks,
-    async (task) => {
-      const messageSent = await sendTagUpdates(
-        task.messageApp,
-        task.chatId,
-        task.updatedRecordsMap
-      );
-
-      if (messageSent) {
-        await User.updateOne(
-          {
-            _id: task.userId,
-            "followedFunctions.functionTag": {
-              $in: [...task.updatedRecordsMap.keys()]
-            } // to avoid duplicate key error
-          },
-          { $set: { "followedFunctions.$[elem].lastUpdate": now } },
-          {
-            arrayFilters: [
-              {
-                "elem.functionTag": { $in: [...task.updatedRecordsMap.keys()] }
-              }
-            ]
-          }
-        );
-      }
-    }
-  );
-}
-
-// There is currently no way to check if a user has been notified of a tag update
-// Resuming an update thus require to force-notify users for all tags updates over the period.
-export async function notifyOrganisationsUpdates(
-  allUpdatedRecords: JORFSearchItem[]
-) {
-  const updatedOrgsWikidataIdSet = new Set<WikidataId>(
-    allUpdatedRecords
-      .flatMap((r) => r.organisations) // all organisations of all records
-      .map((o) => o.wikidata_id) // keep only the id
-      .filter((id): id is WikidataId => !!id) // drop undefined / empty
-  );
-  if (updatedOrgsWikidataIdSet.size == 0) return;
-
-  const updatedOrgsInDb: IOrganisation[] = await Organisation.find({
-    wikidataId: { $in: [...updatedOrgsWikidataIdSet] }
-  }).lean();
-  if (updatedOrgsInDb.length == 0) return;
-
-  const usersFollowingOrganisations: IUser[] = await User.find(
-    {
-      followedOrganisations: {
-        $exists: true,
-        $not: { $size: 0 },
-        $elemMatch: {
-          wikidataId: {
-            $in: updatedOrgsInDb.map((o) => o.wikidataId)
-          }
-        }
-      },
-      status: "active",
-      messageApp: { $in: enabledApps }
-    },
-    {
-      _id: 1,
-      chatId: 1,
-      messageApp: 1,
-      followedOrganisations: { wikidataId: 1, lastUpdate: 1 },
-      schemaVersion: 1
-    }
-  ).lean();
-  if (usersFollowingOrganisations.length == 0) return;
-
-  const orgNameById = new Map<WikidataId, string>(
-    updatedOrgsInDb.map((o) => [o.wikidataId, o.nom])
-  );
-
-  const updatedRecordsWithOrgsInDb = allUpdatedRecords.filter((r) =>
-    r.organisations.some(
-      ({ wikidata_id }) => !!wikidata_id && orgNameById.has(wikidata_id)
-    )
-  );
-  if (updatedRecordsWithOrgsInDb.length == 0) return;
-
-  // Initialize an empty org map to store the categorized records
-  const updatedOrganisationsbyIdMap = new Map<WikidataId, JORFSearchItem[]>();
-
-  // Build the tag map from the updated records
-  updatedRecordsWithOrgsInDb.forEach((item) => {
-    // Iterate through each key in the current record
-    item.organisations.forEach(({ wikidata_id }) => {
-      if (wikidata_id != undefined) {
-        // this should not happen
-        updatedOrganisationsbyIdMap.set(
-          wikidata_id,
-          (updatedOrganisationsbyIdMap.get(wikidata_id) ?? []).concat([item])
-        );
-      }
-    });
-  });
-
-  const now = new Date();
-
-  const userUpdateTasks: NotificationTask<WikidataId>[] = [];
-
-  for (const user of usersFollowingOrganisations) {
-    const newUserOrganisationsUpdates = new Map<WikidataId, JORFSearchItem[]>();
-
-    user.followedOrganisations
-      .filter((orgFollow) =>
-        updatedOrganisationsbyIdMap.has(orgFollow.wikidataId)
-      )
-      .forEach((orgFollow) => {
-        const dateFilteredUserOrgUpdates: JORFSearchItem[] = (
-          updatedOrganisationsbyIdMap.get(orgFollow.wikidataId) ?? []
-        ).filter(
-          (record: JORFSearchItem) =>
-            JORFtoDate(record.source_date).getTime() >
-            orgFollow.lastUpdate.getTime()
-        );
-        if (dateFilteredUserOrgUpdates.length > 0)
-          newUserOrganisationsUpdates.set(
-            orgFollow.wikidataId,
-            dateFilteredUserOrgUpdates
-          );
-      });
-
-    // Calculate the total number of JORFSearchItem in the map
-    let totalUserRecordsCount = 0;
-    newUserOrganisationsUpdates.forEach((items) => {
-      totalUserRecordsCount += items.length;
-    });
-
-    if (totalUserRecordsCount > 0)
-      userUpdateTasks.push({
-        userId: user._id,
-        messageApp: user.messageApp,
-        chatId: user.chatId,
-        updatedRecordsMap: newUserOrganisationsUpdates,
-        recordCount: totalUserRecordsCount
-      });
-  }
-
-  if (userUpdateTasks.length == 0) return;
-
-  await dispatchTasksToMessageApps<WikidataId>(
-    userUpdateTasks,
-    async (task) => {
-      // send follow notification to the user
-      const messageSent = await sendOrganisationUpdate(
-        task.messageApp,
-        task.chatId,
-        task.updatedRecordsMap,
-        orgNameById
-      );
-
-      if (messageSent) {
-        await User.updateOne(
-          {
-            _id: task.userId,
-            "followedOrganisations.wikidataId": {
-              $in: [...task.updatedRecordsMap.keys()]
-            } // to avoid duplicate key error
-          },
-          { $set: { "followedOrganisations.$[elem].lastUpdate": now } },
-          {
-            arrayFilters: [
-              {
-                "elem.wikidataId": {
-                  $in: [...task.updatedRecordsMap.keys()]
-                }
-              }
-            ]
-          }
-        );
-      }
-    }
-  );
-}
-
-export async function notifyPeopleUpdates(updatedRecords: JORFSearchItem[]) {
-  if (updatedRecords.length == 0) return;
-
-  const peopleJSONSet = new Set<string>();
-  updatedRecords.forEach((person) => {
-    peopleJSONSet.add(
-      JSON.stringify({ nom: person.nom, prenom: person.prenom })
-    );
-  });
-
-  // Significantly optimize mongoose request by grouping filters by prenom
-  const byPrenom = [...peopleJSONSet]
-    .map((i) => JSON.parse(i) as { nom: string; prenom: string })
-    .reduce(
-      (acc: Record<string, { nom: string; prenom: string }[]>, person) => {
-        acc[person.prenom] = (acc[person.prenom] ??= []).concat([person]); // Push the object to the array corresponding to its prenom
-        return acc;
-      },
-      {}
-    );
-
-  // Create an array of filter objects
-  const filtersbyPrenom = Object.entries(byPrenom).map(([prenom, arr]) => ({
-    prenom, // This field is used for equality checks, which is index-friendly
-    nom: { $in: arr.map((a) => a.nom) } // This field is used to match any of the "nom" values, which is also index-friendly
-  }));
-
-  const updatedPeopleList: IPeople[] = await People.find({
-    $or: filtersbyPrenom
-  })
-    .collation({ locale: "fr", strength: 2 }) // case-insensitive, no regex
-    .lean();
-  if (updatedPeopleList.length == 0) return;
-
-  // Fetch all users following at least one of the updated People
-  const usersFollowingPeople: IUser[] = await User.find(
-    {
-      followedPeople: {
-        $elemMatch: {
-          peopleId: {
-            $in: updatedPeopleList.map((i) => i._id)
-          }
-        }
-      },
-      status: "active",
-      messageApp: { $in: enabledApps }
-    },
-    {
-      _id: 1,
-      messageApp: 1,
-      chatId: 1,
-      followedPeople: { peopleId: 1, lastUpdate: 1 },
-      schemaVersion: 1
-    }
-  ).lean();
-  if (usersFollowingPeople.length == 0) return;
-
-  const cleanPeopleInfo = updatedPeopleList.map((p) => ({
-    prenom: cleanPeopleName(p.prenom),
-    nom: cleanPeopleName(p.nom)
-  }));
-
-  // Initialize an empty org map to store the categorized records
-  const updatedPeoplebyIdMap = new Map<string, JORFSearchItem[]>();
-
-  // Build the tag map from the updated records
-  updatedRecords.forEach((item) => {
-    const peopleIdx = cleanPeopleInfo.findIndex(
-      (p) =>
-        p.nom === cleanPeopleName(item.nom) &&
-        p.prenom === cleanPeopleName(item.prenom)
-    );
-    if (peopleIdx != -1) {
-      // this should always happen
-      updatedPeoplebyIdMap.set(
-        updatedPeopleList[peopleIdx]._id.toString(),
-        (
-          updatedPeoplebyIdMap.get(
-            updatedPeopleList[peopleIdx]._id.toString()
-          ) ?? []
-        ).concat([item])
-      );
-    }
-  });
-
-  const userUpdateTasks: NotificationTask<string>[] = [];
-
-  const now = new Date();
-
-  const peopleIdMapByStr = new Map<string, Types.ObjectId>();
-  updatedPeopleList.forEach((p) => {
-    peopleIdMapByStr.set(p._id.toString(), p._id);
-  });
-
-  for (const user of usersFollowingPeople) {
-    const newUserPeopleUpdates = new Map<string, JORFSearchItem[]>();
-
-    user.followedPeople
-      .filter((peopleFollow) =>
-        peopleIdMapByStr.has(peopleFollow.peopleId.toString())
-      )
-      .forEach((peopleFollow) => {
-        const dateFilteredUserOrgUpdates: JORFSearchItem[] = (
-          updatedPeoplebyIdMap.get(peopleFollow.peopleId.toString()) ?? []
-        ).filter(
-          (record: JORFSearchItem) =>
-            JORFtoDate(record.source_date).getTime() >
-            peopleFollow.lastUpdate.getTime()
-        );
-        if (dateFilteredUserOrgUpdates.length > 0)
-          newUserPeopleUpdates.set(
-            peopleFollow.peopleId.toString(),
-            dateFilteredUserOrgUpdates
-          );
-      });
-
-    // Calculate the total number of JORFSearchItem in the map
-    let totalUserRecordsCount = 0;
-    newUserPeopleUpdates.forEach((items) => {
-      totalUserRecordsCount += items.length;
-    });
-
-    if (totalUserRecordsCount > 0)
-      userUpdateTasks.push({
-        userId: user._id,
-        messageApp: user.messageApp,
-        chatId: user.chatId,
-        updatedRecordsMap: newUserPeopleUpdates,
-        recordCount: totalUserRecordsCount
-      });
-  }
-
-  if (userUpdateTasks.length == 0) return;
-
-  await dispatchTasksToMessageApps<string>(userUpdateTasks, async (task) => {
-    // send follow notification to the user
-    const messageSent = await sendPeopleUpdate(
-      task.messageApp,
-      task.chatId,
-      task.updatedRecordsMap
-    );
-
-    if (messageSent) {
-      const updatedRecordsPeopleId = [...task.updatedRecordsMap.keys()]
-        .map((idStr) => peopleIdMapByStr.get(idStr))
-        .reduce((tab: Types.ObjectId[], id) => {
-          if (id === undefined) {
-            console.log(
-              "Cannot fetch people id from string during the update of user people follows"
-            );
-            return tab;
-          }
-          return tab.concat(id);
-        }, []);
-
-      await User.updateOne(
-        {
-          _id: task.userId,
-          "followedPeople.peopleId": {
-            $in: updatedRecordsPeopleId
-          } // to avoid duplicate key error
-        },
-        { $set: { "followedPeople.$[elem].lastUpdate": now } },
-        {
-          arrayFilters: [
-            {
-              "elem.peopleId": { $in: updatedRecordsPeopleId }
-            }
-          ]
-        }
-      );
-    }
-  });
-}
-
-export async function notifyNameMentionUpdates(
-  updatedRecords: JORFSearchItem[]
-) {
-  const userFollowingNames: IUser[] = await User.find(
-    {
-      "followedNames.0": { $exists: true },
-      status: "active",
-      messageApp: { $in: enabledApps }
-    },
-    {
-      _id: 1,
-      messageApp: 1,
-      chatId: 1,
-      followedNames: 1,
-      followedPeople: { peopleId: 1, lastUpdate: 1 },
-      schemaVersion: 1
-    }
-  ).lean();
-  if (userFollowingNames.length == 0) return;
-
-  const nameMaps = updatedRecords.reduce(
-    (acc, item: JORFSearchItem) => {
-      const nomPrenom = cleanPeopleName(`${item.nom} ${item.prenom}`);
-      const prenomNom = cleanPeopleName(`${item.prenom} ${item.nom}`);
-
-      const nomPrenomList = acc.nomPrenomMap.get(nomPrenom) ?? []; // existing array or a new one
-      const prenomNomList = acc.prenomNomMap.get(prenomNom) ?? []; // existing array or a new one
-
-      nomPrenomList.push(item);
-      prenomNomList.push(item);
-
-      acc.nomPrenomMap.set(nomPrenom, nomPrenomList); // update the map
-      acc.prenomNomMap.set(prenomNom, prenomNomList); // update the map
-
-      return acc;
-    },
-    {
-      nomPrenomMap: new Map<string, JORFSearchItem[]>(),
-      prenomNomMap: new Map<string, JORFSearchItem[]>()
-    }
-  );
-
-  const now = new Date();
-
-  const userUpdateTasks: NotificationTask<string>[] = [];
-
-  const peopleIdByFollowedNameMap = new Map<string, Types.ObjectId>();
-
-  for (const user of userFollowingNames) {
-    const newUserTagsUpdates = new Map<string, JORFSearchItem[]>();
-
-    for (const followedName of user.followedNames) {
-      const cleanFollowedName = cleanPeopleName(followedName);
-      const mentions =
-        nameMaps.prenomNomMap.get(cleanFollowedName) ??
-        nameMaps.nomPrenomMap.get(cleanFollowedName);
-      if (mentions === undefined || mentions.length == 0) continue;
-
-      const people = await People.findOrCreate({
-        nom: mentions[0].nom,
-        prenom: mentions[0].prenom
-      });
-      peopleIdByFollowedNameMap.set(followedName, people._id);
-      newUserTagsUpdates.set(followedName, mentions);
-    }
-
-    // Calculate the total number of JORFSearchItem in the map
-    let totalUserRecordsCount = 0;
-    newUserTagsUpdates.forEach((items) => {
-      totalUserRecordsCount += items.length;
-    });
-
-    if (totalUserRecordsCount > 0)
-      userUpdateTasks.push({
-        userId: user._id,
-        messageApp: user.messageApp,
-        chatId: user.chatId,
-        updatedRecordsMap: newUserTagsUpdates,
-        recordCount: totalUserRecordsCount
-      });
-  }
-
-  if (userUpdateTasks.length == 0) return;
-
-  await dispatchTasksToMessageApps<string>(userUpdateTasks, async (task) => {
-    await sendNameMentionUpdates(
-      task.messageApp,
-      task.chatId,
-      task.updatedRecordsMap
-    );
-
-    const user = userFollowingNames.find(
-      (u) => u._id.toString() === task.userId.toString()
-    );
-    if (user === undefined) return; // this should not happen
-
-    const peopleIdFollowedByUserStr = user.followedPeople.map((f) =>
-      f.peopleId.toString()
-    );
-
-    const newFollowsIdUnique = [...task.updatedRecordsMap.keys()].reduce(
-      (tab: Types.ObjectId[], idStr) => {
-        const id = peopleIdByFollowedNameMap.get(idStr);
-        if (
-          id !== undefined &&
-          !peopleIdFollowedByUserStr.includes(id.toString())
-        )
-          tab.push(id);
-        return tab;
-      },
-      []
-    );
-
-    await User.updateOne(
-      { _id: user._id },
-      {
-        $pull: {
-          followedNames: {
-            $in: [...task.updatedRecordsMap.keys()]
-          }
-        },
-        $push: {
-          followedPeople: {
-            // add the newFollows to the user followedPeople
-            $each: newFollowsIdUnique.map((id) => ({
-              peopleId: id,
-              lastUpdate: now
-            }))
-          }
-        }
-      }
-    );
-  });
-}
-
-async function sendNameMentionUpdates(
-  messageApp: MessageApp,
-  chatId: IUser["chatId"],
-  updatedRecordMap: Map<string, JORFSearchItem[]>
-): Promise<boolean> {
-  if (updatedRecordMap.size == 0) return true; // no need to send notification if no name mention updates
-
-  // Reverse array change order of records
-  //updatedRecords.reverse();
-
-  const pluralHandler = updatedRecordMap.size > 1 ? "s" : "";
-
-  const markdownLinkEnabled = messageApp !== "Telegram";
-
-  let notification_text = `📢 Nouvelle${pluralHandler} publication${pluralHandler} parmi les noms que vous suivez manuellement:\n\n`;
-
-  const keys = Array.from(updatedRecordMap.keys());
-  const lastKey = keys[keys.length - 1];
-
-  for (const peopleId of updatedRecordMap.keys()) {
-    const nameUpdates = updatedRecordMap.get(peopleId);
-    if (nameUpdates === undefined || nameUpdates.length == 0) {
-      console.log("FollowedName notification update sent with no records");
-      continue;
-    }
-    // Reverse array change order of records
-    // peopleRecords.reverse();
-
-    const prenomNom = nameUpdates[0].prenom + " " + nameUpdates[0].nom;
-
-    const pluralHandler = nameUpdates.length > 1 ? "s" : "";
-    notification_text += `Nouvelle${pluralHandler} publication${pluralHandler} pour ${
-      markdownLinkEnabled
-        ? `[${prenomNom}](${getJORFSearchLinkPeople(prenomNom)})`
-        : `*${prenomNom}*`
-    }\n\n`;
-
-    notification_text += formatSearchResult(nameUpdates, markdownLinkEnabled, {
-      isConfirmation: false,
-      isListing: true,
-      displayName: "no"
-    });
-    notification_text += `Vous suivez maintenant *${prenomNom}* ✅`;
-
-    if (peopleId !== lastKey) notification_text += "====================\n\n";
-  }
-
-  const messageAppsOptionsApp = {
-    ...messageAppsOptions,
-    forceNoKeyboard: messageApp === "WhatsApp"
-  };
-
-  const messageSent = await sendMessage(
-    messageApp,
-    chatId,
-    notification_text,
-    messageAppsOptionsApp
-  );
-  if (messageApp === "WhatsApp")
-    await sendMainMenu(messageApp, chatId, {
-      externalOptions: messageAppsOptions
-    });
-  if (!messageSent) return false;
-
-  await umami.log({ event: "/notification-update-name" });
-  return true;
-}
-
-async function sendPeopleUpdate(
-  messageApp: MessageApp,
-  chatId: IUser["chatId"],
-  updatedRecordMap: Map<string, JORFSearchItem[]>
-) {
-  if (updatedRecordMap.size == 0) return true; // no need to send notification if no name mention updates
-
-  // Reverse array change order of records
-  //updatedRecords.reverse();
-
-  const pluralHandler = updatedRecordMap.size > 1 ? "s" : "";
-
-  const markdownLinkEnabled = messageApp !== "WhatsApp";
-
-  let notification_text = `📢 Nouvelle${pluralHandler} publication${pluralHandler} parmi les personnes que vous suivez :\n\n`;
-
-  const keys = Array.from(updatedRecordMap.keys());
-  const lastKey = keys[keys.length - 1];
-
-  for (const peopleId of updatedRecordMap.keys()) {
-    const peopleRecords = updatedRecordMap.get(peopleId);
-    if (peopleRecords === undefined || peopleRecords.length == 0) {
-      console.log("People notification update sent with no records");
-      continue;
-    }
-    // Reverse array change order of records
-    // peopleRecords.reverse();
-
-    const prenomNom = peopleRecords[0].prenom + " " + peopleRecords[0].nom;
-
-    const pluralHandler = peopleRecords.length > 1 ? "s" : "";
-    notification_text += `Nouvelle${pluralHandler} publication${pluralHandler} pour ${
-      markdownLinkEnabled
-        ? `[${prenomNom}](${getJORFSearchLinkPeople(prenomNom)})`
-        : `*${prenomNom}*`
-    }\n\n`;
-
-    notification_text += formatSearchResult(
-      peopleRecords,
-      markdownLinkEnabled,
-      {
-        isConfirmation: false,
-        isListing: true,
-        displayName: "no"
-      }
-    );
-
-    if (peopleId !== lastKey) notification_text += "====================\n\n";
-  }
-
-  const messageAppsOptionsApp = {
-    ...messageAppsOptions,
-    forceNoKeyboard: messageApp === "WhatsApp"
-  };
-
-  const messageSent = await sendMessage(
-    messageApp,
-    chatId,
-    notification_text,
-    messageAppsOptionsApp
-  );
-  if (messageApp === "WhatsApp")
-    await sendMainMenu(messageApp, chatId, {
-      externalOptions: messageAppsOptions
-    });
-  if (!messageSent) return false;
-
-  await umami.log({ event: "/notification-update-people" });
-  return true;
-}
-
-async function sendOrganisationUpdate(
-  messageApp: MessageApp,
-  chatId: IUser["chatId"],
-  organisationsUpdateRecordsMap: Map<WikidataId, JORFSearchItem[]>,
-  orgNameById: Map<WikidataId, string>
-): Promise<boolean> {
-  if (organisationsUpdateRecordsMap.size == 0) return true;
-
-  let notification_text =
-    "📢 Nouvelles publications parmi les organisations que suivez :\n\n";
-
-  const markdownLinkEnabled = messageApp !== "WhatsApp";
-
-  const keys = Array.from(organisationsUpdateRecordsMap.keys());
-  const lastKey = keys[keys.length - 1];
-
-  for (const orgId of organisationsUpdateRecordsMap.keys()) {
-    const orgName = orgNameById.get(orgId);
-    if (orgName === undefined) {
-      console.log(
-        "Unable to find the name of the organisation with wikidataId " + orgId
-      );
-      continue;
-    }
-    const orgRecords = organisationsUpdateRecordsMap.get(orgId);
-    if (orgRecords === undefined || orgRecords.length == 0) {
-      console.log("Organisation notification update sent with no records");
-      continue;
-    }
-
-    // Reverse array change order of records
-    // orgRecords.reverse();
-
-    const pluralHandler = orgRecords.length > 1 ? "s" : "";
-    notification_text += `Nouvelle${pluralHandler} publication${pluralHandler} pour ${
-      markdownLinkEnabled
-        ? `[${orgName}](${getJORFSearchLinkOrganisation(orgId)})`
-        : `*${orgName}*`
-    }\n\n`;
-
-    notification_text += formatSearchResult(orgRecords, markdownLinkEnabled, {
-      isConfirmation: false,
-      isListing: true,
-      displayName: "all",
-      omitOrganisationNames: true
-    });
-
-    if (orgId !== lastKey) notification_text += DEFAULT_GROUP_SEPARATOR;
-  }
-
-  const messageAppsOptionsApp = {
-    ...messageAppsOptions,
-    forceNoKeyboard: messageApp === "WhatsApp"
-  };
-
-  const messageSent = await sendMessage(
-    messageApp,
-    chatId,
-    notification_text,
-    messageAppsOptionsApp
-  );
-  if (messageApp === "WhatsApp")
-    await sendMainMenu(messageApp, chatId, {
-      externalOptions: messageAppsOptions
-    });
-  if (!messageSent) return false;
-
-  await umami.log({ event: "/notification-update-organisation" });
-  return true;
-}
-
-// We preload the tag keys and values to reduce search time
-const tagValues = Object.values(FunctionTags);
-const tagKeys = Object.keys(FunctionTags);
-
-async function sendTagUpdates(
-  messageApp: MessageApp,
-  chatId: IUser["chatId"],
-  tagMap: Map<FunctionTags, JORFSearchItem[]>
-): Promise<boolean> {
-  // only keep the tags followed by the user
-  const tagList = [...tagMap.keys()];
-
-  if (tagList.length == 0) return true;
-
-  let notification_text =
-    "📢 Nouvelles publications parmi les fonctions que suivez :\n\n";
-
-  const markdownLinkEnabled = messageApp !== "WhatsApp";
-
-  const tagOrder = Array.from(tagMap.keys());
-  const lastTag = tagOrder[tagOrder.length - 1];
-
-  for (const tag of tagOrder) {
-    const tagRecords = tagMap.get(tag);
-    if (tagRecords === undefined || tagRecords.length == 0) {
-      console.log("Tag notification update sent with no records");
-      continue;
-    }
-    const tagKey = tagKeys[tagValues.indexOf(tag)];
-
-    // Reverse array change order of records
-    // updatedRecords.reverse();
-
-    const pluralHandler = tagRecords.length > 1 ? "s" : "";
-    notification_text += `Nouvelle${pluralHandler} publication${pluralHandler} pour la fonction ${
-      markdownLinkEnabled
-        ? `[${tagKey}](${getJORFSearchLinkFunctionTag(tag)})`
-        : `*${tagKey}*`
-    }\n\n`;
-
-    const groupingConfig = functionTagGroupingStrategies[tag];
-    let handledByGrouping = false;
-
-    if (groupingConfig) {
-      const groupedMap = groupRecordsBy(tagRecords, groupingConfig);
-      const orderedEntries = orderGroupedEntries(
-        groupedMap,
-        groupingConfig.sortGroupIds
-      ).filter(([, records]) => records.length > 0);
-
-      if (orderedEntries.length > 0) {
-        handledByGrouping = true;
-
-        orderedEntries.forEach(([groupId, groupRecords], groupIndex) => {
-          const groupTitle =
-            groupingConfig.formatGroupTitle?.({
-              groupId,
-              markdownLinkEnabled
-            }) ?? `👉 ${groupId}\n\n`;
-
-          notification_text += groupTitle;
-
-          notification_text += formatSearchResult(
-            groupRecords,
-            markdownLinkEnabled,
-            {
-              isConfirmation: false,
-              isListing: true,
-              displayName: "all",
-              omitOrganisationNames:
-                groupingConfig.omitOrganisationNames ?? false
-            }
-          );
-
-          const isLastGroup = groupIndex === orderedEntries.length - 1;
-          if (!isLastGroup) notification_text += DEFAULT_SUBGROUP_SEPARATOR;
-          else if (tag !== lastTag) notification_text += DEFAULT_GROUP_SEPARATOR;
-        });
-      }
-    }
-
-    if (handledByGrouping) continue;
-
-    notification_text += formatSearchResult(tagRecords, markdownLinkEnabled, {
-      isConfirmation: false,
-      isListing: true,
-      displayName: "all",
-      omitOrganisationNames: groupingConfig?.omitOrganisationNames ?? false
-    });
-
-    if (tag !== lastTag) notification_text += DEFAULT_GROUP_SEPARATOR;
-  }
-
-  const messageAppsOptionsApp = {
-    ...messageAppsOptions,
-    forceNoKeyboard: messageApp === "WhatsApp"
-  };
-
-  const messageSent = await sendMessage(
-    messageApp,
-    chatId,
-    notification_text,
-    messageAppsOptionsApp
-  );
-  if (messageApp === "WhatsApp")
-    await sendMainMenu(messageApp, chatId, {
-      externalOptions: messageAppsOptions
-    });
-  if (!messageSent) return false;
-
-  await umami.log({ event: "/notification-update-function" });
-  return true;
-}
-
-await (async () => {
-  // Connect to DB
-  await mongodbConnect();
-
-  // the currentDate is today
-  const currentDate = new Date();
-  const startDate = new Date(
-    currentDate.getFullYear(),
-    currentDate.getMonth(),
-    currentDate.getDate() - SHIFT_DAYS
-  );
-  startDate.setHours(0, 0, 0, 0);
-  // Fetch all records from the start date
-  const JORFAllRecordsFromDate = await getJORFRecordsFromDate(startDate);
-
-  if (JORFAllRecordsFromDate.length > 0) {
-    // Send notifications to users on followed people
-    await notifyPeopleUpdates(JORFAllRecordsFromDate);
-
-    // Send notifications to users on followed names
-    await notifyNameMentionUpdates(JORFAllRecordsFromDate);
-
-    // Send notifications to users on followed functions
-    await notifyFunctionTagsUpdates(JORFAllRecordsFromDate);
-
-    // Send notifications to users on followed organisations
-    await notifyOrganisationsUpdates(JORFAllRecordsFromDate);
-  }
-
-  await umami.log({ event: "/notification-process-completed" });
-
-  process.exit(0);
-})();
